+++ internal/version/version_test.go
@@ -3,10 +3,10 @@ package version
 import "testing"
 
 func TestString(t *testing.T) {
-	if Version != "0.4.1" {
+	if Version != "0.4.2" {
 		t.Fatalf("Version=%s", Version)
 	}
-	if String() != "p2pirc 0.4.1" {
+	if String() != "p2pirc 0.4.2" {
 		t.Fatalf("String=%s", String())
 	}
 }